print anything

+++ 0x10-variadic_functions/3-print_all.c
@@ -32,7 +32,7 @@ void print_all(const char * const format, ...)
 				printf("%d", va_arg(valist, int)), k = 1;
 				break;
 			case 'f':
-				printf("%f", va_arg(valist, int)), k = 1;
+				printf("%f", va_arg(valist, double, k = 1;
 				break;
 			case 's':
 				str = va_arg(valist, char *), k = 1;
